requirements.txt
Flask
flask_uploads
pillow
pytesseract
requests
